+++ digitaltwin_dataspace/components/base.py
