+++ digitaltwin_dataspace/data/table.py
