+++ digitaltwin_dataspace/data/write.py
