+++ digitaltwin_dataspace/utils.py
